requirements.txt
PyMuPDF==1.24.9
pdfplumber==0.8.0
Pillow==10.0.0
pandas==2.2
